server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
binwalk==2.3.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cssselect2==0.9.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.134.0
filelock==3.13.1
fonttools==4.61.1
fs==2.4.16
gyp==0.1
h11==0.16.0
html5lib==1.1
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyboard==0.13.5
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
logging==0.4.9.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
Markdown==3.10.2
markdown-it-py==3.0.0
markdownify==1.2.2
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mpmath==0.0.0
netaddr==0.8.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.5
packaging==24.0
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==3.0
pycups==2.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
pygame==2.5.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pymongo==4.12.1
PyMuPDF==1.27.1
PyOpenGL==3.1.7
pyparsing==3.1.1
pypdf==6.7.3
pyphen==0.17.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-sip==13.6.0
pyqtgraph==0.13.4
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.22
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.4.10
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
starlette==0.52.1
sympy==1.12
systemd-python==235
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
userpath==1.9.1
uvicorn==0.41.0
vboxapi==1.0
virtualenv==20.25.0+ds
wadllib==1.3.6
weasyprint==68.1
webencodings==0.5.1
xdg==5
xkit==0.0.0
zombie-imp==0.0.2
zopfli==0.4.1
